# openapi.yaml
openapi: 3.0.0
info:
  title: Banking API
  version: 1.0.0
  description: >-
    This is a sample banking API that I used to learn a bunch of tools, such as
    Kotlin, Spring, Gradle and Swagger. This is NOT inspired by any real world
    applications and is most definitely not production-ready (nor will it ever
    be), I'm simply using it as a learning tool.
  contact:
    name: Lucas de Ataides
    email: ldatb@icloud.com
    url: https://ldatb.com/
  license:
    name: MIT
    url: https://opensource.org/license/mit/
...
paths:
...
  /account:
    summary: Access to accounts
    post:
      summary: Add a new account
      operationId: addAccount
      tags:
        - account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreationRequest'
      responses:
        '201':
          description: Account was created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCreationResponse'
        '400':
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
    get:
      summary: Gets basic information of the account
      operationId: getAccountByNumber
      tags:
        - account
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          description: Missing JWT in request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedErrorResponse'
    delete:
      summary: Deletes an account
      operationId: deleteAccount
      tags:
        - account
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Account was deleted
        '401':
          description: Missing JWT in request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedErrorResponse'
...
components:
  schemas:
...
    AccountCreationRequest:
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        password:
          type: string
          example: Super_Secret_Password
      required:
        - firstName
        - lastName
        - password
...
    AccountResponse:
      type: object
      properties:
        error:
          type: boolean
          example: false
        data:
          type: object
          properties:
            accountLogin:
              type: string
              example: 'someaccountlogin'
            transferKey:
              description: Instead of relying on the account login for transactions, which can lead to security issues, a different value is used for transactions
              type: string
              example: '550e8400-e29b-41d4-a716-446655440000'
            firstName:
              type: string
              example: John
            lastName:
              type: string
              example: Doe
            balance:
              description: 'Balance in account. The last 2 numbers represent the cents'
              type: number
              format: int64
              example: 6713
            currency:
              description: 'The currency of the balance'
              type: string
              example: 'usd'
    AccountCreationResponse:
      type: object
      properties:
        error:
          type: boolean
          example: false
        data:
          type: object
          properties:
            accountLogin:
              type: string
              example: 'someaccountlogin'
            transferKey:
              description: Instead of relying on the account login for transactions, which can lead to security issues, a different value is used for transactions
              type: string
              example: '550e8400-e29b-41d4-a716-446655440000'
            secretToken:
              description: This token will be used in transactions
              type: integer
              example: 123456
            firstName:
              type: string
              example: John
            lastName:
              type: string
              example: Doe
...
    GenericErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Invalid request
        details:
          type: string
          example: Error details
        exception:
          type: string
          example: HttpMessageNotReadableException
    UnauthorizedErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Missing access token in request
        details:
          type: string
          example: Error details
        exception:
          type: string
          example: HttpMessageNotReadableException
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer